face-detector/requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
Pillow>=10.0,<12.0

# Optional: install facenet-pytorch if you want the MTCNN backend.
# facenet-pytorch>=2.6.0
